backend/requirements.txt
# PanelLab backend deps

# Web
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
cryptography>=41.0.0

# Config
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
typing_extensions>=4.0.0

# Auth
PyJWT>=2.8.0
bcrypt>=4.0.0

# System monitor (stage 2)
psutil>=5.9.0
paramiko>=3.0.0